Splitting lua files
