+++ config/neovim/treesitter.lua
@@ -1,0 +1,1 @@
+--
